feat: add more pretty icons and background

+++ app/src/main/java/com/onopry/whac_a_mole/Utills.kt
@@ -10,5 +10,5 @@ const val MOLE_LIFE_TIME = 1000 // milliseconds
 const val GAME_TIME = 100000L
 const val GAME_TICK = 200L
 
-fun Int.toDp(): Int = (this / Resources.getSystem().displayMetrics.density).toInt()
-fun Int.toPx() = this * Resources.getSystem().displayMetrics.density
+fun Float.toDp(): Float = (this / Resources.getSystem().displayMetrics.density)
+fun Float.toPx() = this * Resources.getSystem().displayMetrics.density
